Move Logger.measure() and AXElement.depth() to their own type def
