+++ maestro-ios-xctest-runner/maestro-driver-iosUITests/Routes/Extensions/Logger.swift
@@ -1,0 +1,17 @@
+
+import Foundation
+import os
+
+extension Logger {
+    func measure<T>(message: String, _ block: () throws -> T) rethrows -> T {
+        let start = Date()
+        info("\(message) - start")
+
+        let result = try block()
+
+        let duration = Date().timeIntervalSince(start)
+        info("\(message) - duration \(duration)")
+
+        return result
+    }
+}
